--- .pre-commit-config.yaml
@@ -1,23 +1,0 @@
-repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.5.0
-    hooks:
-    -   id: check-yaml
-    -   id: end-of-file-fixer
-#     # -   id: trailing-whitespace
-# # -   repo: https://github.com/hhatto/autopep8
-# #     rev: v2.0.4
-# #     hooks:
-# #     - id: autopep8
-# #       args: [--in-place, --recursive]
-
-# -   repo: https://github.com/psf/black
-#     rev: 23.11.0
-#     hooks:
-#     - id: black
-        # args: [--line-length=88]
-# -   repo: https://github.com/PyCQA/flake8
-#     rev: 6.1.0
-#     hooks:
-#     - id: flake8
-#       args: [--max-line-length=88]
