+++ efold/models/factory.py
@@ -1,5 +1,5 @@
 from .transformer import Transformer
-from .evoformer import Evoformer
+from .efold import eFold
 from .cnn import CNN
 from .ribonanza import Ribonanza
 from .unet import U_Net
@@ -8,8 +8,8 @@ from .unet import U_Net
 def create_model(model: str, *args, **kwargs):
     if model == "transformer":
         return Transformer(*args, **kwargs)
-    if model == "evoformer":
-        return Evoformer(*args, **kwargs)
+    if model == "efold":
+        return eFold(*args, **kwargs)
     if model == "cnn":
         return CNN(*args, **kwargs)
     if model == "unet":
